swiftgrade-api/assessments/serializers/assessment_for_teacher/unassign_assessment_serializer.py
from rest_framework import serializers

from assessments.services import AssessmentService


class UnassignAssessmentSerializer(serializers.Serializer):
    group_id = serializers.IntegerField(required=True)
