swiftgrade-api/requirements.txt
aiohttp==3.7.4.post0
amqp==2.6.1
antlr4-python3-runtime==4.7.1
appnope==0.1.0
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
billiard==3.6.3.0
boto==2.49.0
boto3==1.12.21
botocore==1.15.49
cachetools==4.2.0
celery==4.4.7
certifi==2020.6.20
cffi==1.15.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
cycler==0.10.0
decorator==4.4.2
Django==2.2
django-admin-rangefilter==0.5.4
django-bulk-update==2.2.0
django-cors-headers==3.2.1
django-dotenv==1.4.2
django-dynamic-preferences==1.7
django-extensions==2.2.8
django-filter==2.0.0
django-hashid-field==2.1.6
django-jet==1.0.8
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
docutils==0.15.2
drf-friendly-errors==0.13
django-celery-beat==2.2.0
drf-yasg==1.17.0
Faker==4.0.2
fpdf==1.7.2
google-auth==1.24.0
gunicorn==20.0.4
hashids==1.3.1
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
imutils==0.5.3
inflection==0.5.1
ipython==7.23.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
kiwisolver==1.3.1
kombu==4.6.11
MarkupSafe==1.1.1
matplotlib==3.4.2
matplotlib-inline==0.1.3
more-itertools==8.7.0
mpmath==1.1.0
multidict==5.1.0
nameparser==1.0.6
networkx==2.5.1
numpy==1.18.3
opencv-python==3.4.4.19
packaging==20.4
pandas==1.0.3
parso==0.7.1
persisting-theory==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.7
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.19.1
pycparser==2.20
Pygments==2.7.1
PyJWT==2.3.0
PyMuPDF==1.17.0
pyparsing==2.4.7
PyPDF2==1.26.0
pytexit==0.3.3
python-barcode==0.11.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pyzbar==0.1.8
redis==3.3.0
requests==2.24.0
rest-condition==1.0.3
rsa==4.7
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
scikit-image==0.16.2
scipy==1.7.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
sympy==1.5.1
text-unidecode==1.3
textdistance==4.2.0
traitlets==5.0.4
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
whitenoise==5.0.1
xxhash==2.0.0
xlsxwriter==3.0.3
yarl==1.6.3
zipp==3.3.0
